+++ asdf_astropy/converters/coordinates/representation.py
@@ -19,6 +19,21 @@ class RepresentationConverter(Converter):
         "astropy.coordinates.representation.UnitSphericalCosLatDifferential",
         "astropy.coordinates.representation.UnitSphericalDifferential",
         "astropy.coordinates.representation.UnitSphericalRepresentation",
+        # classes were moved in https://github.com/astropy/astropy/pull/14792
+        "astropy.coordinates.representation.cartesian.CartesianDifferential",
+        "astropy.coordinates.representation.cartesian.CartesianRepresentation",
+        "astropy.coordinates.representation.cylindrical.CylindricalDifferential",
+        "astropy.coordinates.representation.cylindrical.CylindricalRepresentation",
+        "astropy.coordinates.representation.spherical.PhysicsSphericalDifferential",
+        "astropy.coordinates.representation.spherical.PhysicsSphericalRepresentation",
+        "astropy.coordinates.representation.spherical.RadialDifferential",
+        "astropy.coordinates.representation.spherical.RadialRepresentation",
+        "astropy.coordinates.representation.spherical.SphericalCosLatDifferential",
+        "astropy.coordinates.representation.spherical.SphericalDifferential",
+        "astropy.coordinates.representation.spherical.SphericalRepresentation",
+        "astropy.coordinates.representation.spherical.UnitSphericalRepresentation",
+        "astropy.coordinates.representation.spherical.UnitSphericalDifferential",
+        "astropy.coordinates.representation.spherical.UnitSphericalCosLatDifferential",
     ]
 
     def to_yaml_tree(self, obj, tag, ctx):