+++ CHANGES.rst
@@ -2,6 +2,7 @@
 ------------------
 
 - Drop support for Python 3.8 in accordance with NEP 29. [#180]
+- Update ``RepresentationConverter`` for new class paths in astropy [#181]
 
 0.4.0 (2023-03-20)
 ------------------